Add process doppelganging injection support via a shellcode. Solve #332
Provides the shellcode and the source code use to generate it.

Allows to inject a binary into the guest's filesystem and execute it. It
has some limitations due to the technique itself as well as the current
implementation. For example certain programs fail to locate some DLLs.

Also, you need to have write permission on the guest process used as a
cover otherwise the call to `CreateFileTransacted` will fail.