Added SMC traps for ARM64 platforms
This patch adds an preprocessor to differentiate x86 and ARM64.
This patch further adds the function smc_cb which is similar to the int3_cb.
smc_cb also helps us to emulate single-stepping:

As Xen does not support single-stepping on ARM, yet, we emulate the
single-stepping behavior by using two logically adjacent SMC instructions
placed in two different views; the first SMC instruction is placed instead
of the original instruction in the execute-view. The second SMC instruction
follows the original instruction in the step-view. By switching to the
step-view right after the first SMC invocation, the system will trap
immediately after executing the original instruction. In this way, this allows
the monitor to switch back to the exec-view and continue execution.

Because this patch uses inject_trap_breakpoint in the code path,
we renamed the function to inject_trap_sw.

This patch need the libvmi commit: 69b3198 to build.