Process doppelganging injection support (#444)
* [In progress..] Universal function to set the stack

* [In progress..] Universal function to set the stack

* [In progress..] Universal function to set the stack

* Universal function to set the stack (64 bit only)

* Fixes astyle

* [In progress..] Exec shellcode : add VirtualAlloc support

* [In progress..] Exec shellcode : write SC to memory

* Revert "[In progress..] Exec shellcode : write SC to memory"

This reverts commit b20ff6772551bf283c1f70ff2ab5f2f50ca83721.

* Revert "[In progress..] Exec shellcode : add VirtualAlloc support"

This reverts commit 11f60cc61660752877f832ba8fb1ed386f69fe58.

* Universal funciton to set the stack (32 bit)

* Fixes astyle

* Add process doppelganging injection support via a shellcode. Solve #332

Provides the shellcode and the source code use to generate it.

Allows to inject a binary into the guest's filesystem and execute it. It
has some limitations due to the technique itself as well as the current
implementation. For example certain programs fail to locate some DLLs.

Also, you need to have write permission on the guest process used as a
cover otherwise the call to `CreateFileTransacted` will fail.

* Add process doppelganging injection support via a shellcode. Solve #332

Provides the shellcode and the source code use to generate it.

Allows to inject a binary into the guest's filesystem and execute it. It
has some limitations due to the technique itself as well as the current
implementation. For example certain programs fail to locate some DLLs.

Also, you need to have write permission on the guest process used as a
cover otherwise the call to `CreateFileTransacted` will fail.

* Adds a more detail comment

* Fixes astyle for shellcode source code

* Fixes astyle for shellcode source code

* Adds precision to help message

* Removes unused structure

* Fix infinite loop when additional breakpoint is never hit

* Check Windows build before setting additional breakpoint

* Move build date check

* Add DRAKVUF licence