Painless Xen 4.12 builds + minor C++ improvements (#480)
* Fix Xen 4.12-staging build

The code needs to link with -lxenforeignmemory and -lxencall for
recent versions of Xen. Check for those libraries (and link with
them if found) - but still allow them to be optional.

Signed-off-by: Razvan Cojocaru <rcojocaru@bitdefender.com>

* Un-constify pass-by-value parameters to drakvuf_c::drakvuf_c()

Signed-off-by: Razvan Cojocaru <rcojocaru@bitdefender.com>

* Remove unnecessary this-> qualifiers

Signed-off-by: Razvan Cojocaru <rcojocaru@bitdefender.com>

* Replace NULL with nullptr in C++ code

Signed-off-by: Razvan Cojocaru <rcojocaru@bitdefender.com>

* Use C++ initializer lists and in-class initialization

Signed-off-by: Razvan Cojocaru <rcojocaru@bitdefender.com>

* Remove #include <stdbool.h>

Booleans are first-class citizens in C++ (supported directly by
language keywords), so we don't need to #include <stdbool.h> in
C++ code.

Signed-off-by: Razvan Cojocaru <rcojocaru@bitdefender.com>

* Use nullptr instead of NULL in plugins.h

Signed-off-by: Razvan Cojocaru <rcojocaru@bitdefender.com>

* Use the newer virtual ~destructor() = default; form

For base classes, starting with C++11 we can say:

class Base {
public:
	virtual ~Base() = default;
	// Instead of: virtual ~Base() {}
};

It's succint and clearer in what it is trying to achieve.

Signed-off-by: Razvan Cojocaru <rcojocaru@bitdefender.com>

* Don't throw int

Use std::runtime_error with an informative message instead.

Signed-off-by: Razvan Cojocaru <rcojocaru@bitdefender.com>

* Remove redundant this-> from plugins.{h,cpp}

Signed-off-by: Razvan Cojocaru <rcojocaru@bitdefender.com>

* Minor cleanup in ~drakvuf_plugins()

for ( int i = start; ... ) is idiomatic C++ rather than
int i; for ( i = start; ... ). Also, delete nullptr; is guaranteed
to work properly.

Signed-off-by: Razvan Cojocaru <rcojocaru@bitdefender.com>

* Replace NULL with nullptr in main.cpp

Signed-off-by: Razvan Cojocaru <rcojocaru@bitdefender.com>

* Minor C++ cleanup in main.cpp

Use true and false instead of 0 and 1 for bools, use
for( int i = value; ... ) instead of int i; for ( i = value; ... )

Signed-off-by: Razvan Cojocaru <rcojocaru@bitdefender.com>

* Bugfix: don't start out wait_for_process with random stack value

Signed-off-by: Razvan Cojocaru <rcojocaru@bitdefender.com>

* Bugfix: use strtol() to determine numerical value

The current code does sscanf(input, "%u", &_domID); to figure out
if input is a number, but that fails for domain names that start
with a number (such as "4xenwin7", where it puts 4 into _domID).

Use strtol() to determine whether the _whole_ input string is a
number or not.

Signed-off-by: Razvan Cojocaru <rcojocaru@bitdefender.com>