procdump: make more robust version
NOTE This is a part of big re-factoring. In the meantime I separate the efforts
into "procdump2" plug-in. It would replace the original "procdump" on ready.

In thhe current implementation of "procdump" we copy process address
space with RtlCopyMemoryNonTemporal at MmCleanProcessAddressSpace. This
assumes the good knowledge of VAD attributes like type, MemCommit,
CommitCharge, etc. While the minor error in understanding this
attributes result in BSODs. Ever more due to this implementation it is
impossible for us to dump running processes.

New implementation (see "procdump2") uses less knowledge of VADs. Only
start address and length are essential. MmCopyVirtualMemory is used to
copy process address space from the context of other process. Instead of
BSODs this function just return fail status if some memory region is
inaccessible any more. And this allow us to dump running processes. This
feature is used to dump injected process while stopping the plug-in.