filedelete: re-write plugin to clarify model
Before refactoring multiple hooks on NtClose been set for each
step in injection chain. Thus the model been difficult to understand.
The other drawback was that the order of callbacks been called was
crucial.

With new model the single hook on NtClose left. And every task get
it's state machine. Thus the model been clarified a lot.