filedelete: re-write plugin to clarify model (#1393)
* filedelete: re-write plugin to clarify model

Before refactoring multiple hooks on NtClose been set for each
step in injection chain. Thus the model been difficult to understand.
The other drawback was that the order of callbacks been called was
crucial.

With new model the single hook on NtClose left. And every task get
it's state machine. Thus the model been clarified a lot.

* filedelete: fix typo

* fileextractor: move into separate plug-in

* fileextractor: a little bit of refactoring

* fileextractor: fix file perms for CodeQL

* fileextractor: fix review issues for return hook

* plugin_ex: fix type error

The error results in UndefinedBehaviorSanitizer panic on type change.

With additional debug prints:
```
[createfile_cb_impl:166] '19createfile_result_t'
[createReturnHook:599] '19createfile_result_t'
[createReturnHook:606] '12PluginResult'
```

* fileextractor: fix astyle

Co-authored-by: Sergey Kovalev <skovalev@ptsecurity.com>