Workaround possible PF injection in case of vmi_request_page_fault failure.
From drakvuf stderr:

	[USERHOOK] Export info not accessible, page fault 774c0000
	VMI_ERROR: xen_request_page_fault error -1 injecting page fault exception
	[USERHOOK] Failed to request page fault for DTB 63cbb000, address 774c1000
	[USERHOOK] Entered system service handler
	[USERHOOK] Not suppressing service exception - not our fault

It looks like the `vmi_request_page_fault` returned response that it could not
inject a page fault, but it did. The situation happened only once.

I could not understand from the XEN source code how this could happen.
The hypervisor either returns an error or does what it was asked to do.